Fix set unlocking screen visibility after updating keys
