+++ tella-locking-ui/src/main/java/com/hzontal/tella_locking_ui/Constants.kt
@@ -3,3 +3,5 @@ package com.hzontal.tella_locking_ui
 const val IS_FROM_SETTINGS = "IS_FROM_SETTINGS"
 
 const val ONBOARDING_CLASS_NAME = "rs.readahead.washington.mobile.views.activity.onboarding.OnBoardingActivity"
+
+const val FINISH_ACTIVITY_REQUEST_CODE = 123
