+++ tella-locking-ui/src/main/java/com/hzontal/tella_locking_ui/common/BaseActivity.kt
@@ -3,10 +3,7 @@ package com.hzontal.tella_locking_ui.common
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
-import androidx.activity.result.ActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import com.hzontal.tella_locking_ui.IS_FROM_SETTINGS
 import com.hzontal.tella_locking_ui.ONBOARDING_CLASS_NAME
@@ -18,33 +15,17 @@ import org.hzontal.tella.keys.config.UnlockConfig
 import org.hzontal.tella.keys.config.UnlockRegistry
 import org.hzontal.tella.keys.key.MainKey
 import timber.log.Timber
-import java.util.concurrent.Executors
-import java.util.concurrent.TimeUnit
 
 open class BaseActivity : AppCompatActivity() {
     protected val isFromSettings by lazy { intent.getBooleanExtra(IS_FROM_SETTINGS, false) }
     private var isConfirmSettingsUpdate: Boolean = false
     protected val config: UnlockConfig by lazy { TellaKeysUI.getUnlockRegistry().getActiveConfig(this) }
     protected val registry: UnlockRegistry by lazy { TellaKeysUI.getUnlockRegistry() }
-    var startForResult = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result: ActivityResult ->
-        if (result.resultCode == Activity.RESULT_OK) { finish()}
-    }
-    var finishActivity = MutableLiveData<Boolean>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         Timber.d("** %s: %s **", javaClass, "onCreate()")
         super.onCreate(savedInstanceState)
         overridePendingTransition(R.anim.`in`, R.anim.out)
-        CommonStates.finishUpdateActivity.observe(this, Observer {isFinished->
-            if (isFinished) {
-                TellaKeysUI.getCredentialsCallback().onUpdateUnlocking()
-                finish()
-            }
-        })
-    }
-
-    override fun onBackPressed() {
-        finish()
     }
 
     override fun onResume() {
@@ -79,14 +60,11 @@ open class BaseActivity : AppCompatActivity() {
 
     protected fun onSuccessConfirmUnlock() {
         if (isConfirmSettingsUpdate) {
-            val dialog = SuccessUpdateDialog()
-            dialog.show(supportFragmentManager, "SUCCESS_DIALOG")
-            val executor = Executors.newSingleThreadScheduledExecutor()
-            val hideDialog = Runnable {
-                CommonStates.finishUpdateActivity.postValue(true)
-                dialog.dismiss()
-            }
-            executor.schedule(hideDialog, 3, TimeUnit.SECONDS);
+            TellaKeysUI.getCredentialsCallback().onUpdateUnlocking()
+            val intent = Intent(this, SuccessUpdateDialog::class.java)
+            setResult(Activity.RESULT_OK)
+            startActivity(intent)
+            finish()
         } else {
             startActivity(Intent(this, ConfirmCredentialsActivity::class.java))
             finishAffinity()
@@ -100,6 +78,5 @@ open class BaseActivity : AppCompatActivity() {
         } else {
             MainKey.generate()
         }
-
     }
 }
