+++ tella-locking-ui/src/main/java/com/hzontal/tella_locking_ui/ui/pattern/PatternSetConfirmActivity.kt
@@ -70,9 +70,4 @@ class PatternSetConfirmActivity : SetPatternActivity() {
         })
     }
 
-    override fun onDestroy() {
-        super.onDestroy()
-        CommonStates.finishUpdateActivity.postValue(false)
-    }
-
 }
