+++ tella-locking-ui/src/main/java/com/hzontal/tella_locking_ui/ui/pin/SetPinActivity.kt
@@ -1,8 +1,9 @@
 package com.hzontal.tella_locking_ui.ui.pin
 
+import android.app.Activity
 import android.content.Intent
+import com.hzontal.tella_locking_ui.FINISH_ACTIVITY_REQUEST_CODE
 import com.hzontal.tella_locking_ui.R
-import com.hzontal.tella_locking_ui.common.CommonStates
 import com.hzontal.tella_locking_ui.ui.pin.base.BasePinActivity
 
 const val CONFIRM_PIN = "confirm_pin"
@@ -10,17 +11,17 @@ class SetPinActivity : BasePinActivity() {
 
     override fun onSuccessSetPin(pin: String?) {
         val intent = Intent(this, ConfirmPinActivity::class.java)
-        intent.putExtra(CONFIRM_PIN,pin)
-        startActivity(intent)
-        overridePendingTransition(R.anim.`in`,R.anim.out)
+        intent.putExtra(CONFIRM_PIN, pin)
+        intent.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP
+        startActivityForResult(intent, FINISH_ACTIVITY_REQUEST_CODE)
+        overridePendingTransition(R.anim.`in`, R.anim.out)
     }
 
     override fun onFailureSetPin(error: String) {
     }
 
-    override fun onDestroy() {
-        super.onDestroy()
-        CommonStates.finishUpdateActivity.postValue(false)
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (requestCode == FINISH_ACTIVITY_REQUEST_CODE && resultCode == Activity.RESULT_OK) finish()
     }
-
 }
